Added two more web pages
Changes pending: Approve button in company-home page is not working. Potential error in future: Apply button in worker-home page also might not work because since the template consists of only one button it has been given an id and not a class. Hence the javascript functionality works for the worker-home apply button as it belongs to an id, and the approvebutton in company home doesnt work because it belongs to a class and there are two of them, inside the template.